% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{board_folder}
\alias{board_folder}
\title{Create a folder-based storage backend}
\usage{
board_folder(path)
}
\arguments{
\item{path}{Path to the root directory for workflow storage.}
}
\description{
Create a folder-based storage backend
}
