% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coal}
\alias{coal}
\title{Coalesce - return first non-NULL value}
\usage{
coal(...)
}
\arguments{
\item{...}{Values to check}
}
\value{
First non-NULL value
}
\description{
Coalesce - return first non-NULL value
}
\keyword{internal}
