% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_time_ago}
\alias{format_time_ago}
\title{Format time ago}
\usage{
format_time_ago(time)
}
\arguments{
\item{time}{POSIXct or character timestamp}
}
\value{
Character string like "2 hours ago", "3 days ago"
}
\description{
Converts a timestamp to a human-readable "time ago" string.
}
\keyword{internal}
