% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_initials}
\alias{get_initials}
\title{Get initials from username}
\usage{
get_initials(username)
}
\arguments{
\item{username}{Character string}
}
\value{
1-2 character initials
}
\description{
Get initials from username
}
\keyword{internal}
