% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{manage_session}
\alias{manage_session}
\alias{manage_session_server}
\alias{manage_session_ui}
\title{Session management}
\usage{
manage_session(server = manage_session_server, ui = manage_session_ui)

manage_session_server(id, board, ...)

manage_session_ui(id, board)
}
\arguments{
\item{server, ui}{Server/UI for the plugin module}

\item{id}{Namespace ID}

\item{board}{The initial \code{board} object}

\item{...}{Extra arguments passed from parent scope}
}
\value{
See \code{\link[blockr.core:preserve_board]{blockr.core::preserve_board()}}.
}
\description{
Provides a \code{\link[blockr.core:preserve_board]{blockr.core::preserve_board()}} using a \code{\link[pins:board_folder]{pins::board_local()}} as
storage backend. The \code{\link[=blockr_option]{blockr_option()}} \code{session_mgmt_backend} can be set
to swap out this default pin board for another.
}
