% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-extension.R
\name{new_session_extension}
\alias{new_session_extension}
\title{Session extension}
\usage{
new_session_extension()
}
\value{
A navbar provider object
}
\description{
Adds workflow management UI to the blockr.dock navbar (save, load,
workflows list, history). Unlike dock extensions, this does not create
a dock panel - it only provides navbar content.
}
\examples{
\dontrun{
library(blockr.dock)
library(blockr.session)

serve(
  new_dock_board(
    extensions = list(
      blockr.dag::new_dag_extension(),
      new_session_extension()
    )
  )
)
}

}
