% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{rack_file}
\alias{rack_file}
\title{Get the local file path for a workflow version}
\usage{
rack_file(id, backend, ...)
}
\arguments{
\item{id}{A rack ID.}

\item{backend}{A storage backend.}

\item{...}{Additional arguments.}
}
\description{
Get the local file path for a workflow version
}
\keyword{internal}
