% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-right-server.R
\name{session_navbar_right_server}
\alias{session_navbar_right_server}
\title{Session navbar right slot server}
\usage{
session_navbar_right_server(
  id,
  board = NULL,
  dock = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Namespace ID}

\item{board}{Reactive values object (unused, for interface consistency)}

\item{dock}{Dock object (unused, for interface consistency)}

\item{session}{Shiny session}
}
\description{
Renders user avatar with initials.
}
