% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-right-ui.R
\name{session_navbar_right_ui}
\alias{session_navbar_right_ui}
\title{Session navbar right slot UI}
\usage{
session_navbar_right_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
Shiny UI tagList
}
\description{
Creates the right side navbar content for blockr.dock integration:
\itemize{
\item User avatar with initials
}
}
