% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar-right-ui.R
\name{user_avatar_ui}
\alias{user_avatar_ui}
\title{User avatar UI helper}
\usage{
user_avatar_ui(initials = "U")
}
\arguments{
\item{initials}{User initials (1-2 characters)}
}
\value{
Shiny UI element
}
\description{
User avatar UI helper
}
\keyword{internal}
