% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.session-package}
\alias{blockr.session}
\alias{blockr.session-package}
\title{blockr.session: Session management for blockr}
\description{
Provides a "preserve_board" module for blockr apps to save and manage projects and restore existing boards.
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

}
\keyword{internal}
