% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_workflows}
\alias{list_workflows}
\title{List all workflows}
\usage{
list_workflows(backend)
}
\arguments{
\item{backend}{A pins board}
}
\value{
Data frame with columns: name, created, time_ago
}
\description{
Returns a data frame of all saved workflows with metadata.
}
\keyword{internal}
